+++ packages/view/src/types/GitHubUser.ts
@@ -1,0 +1,5 @@
+export type GitHubUser = {
+  id: string;
+  names: string[];
+  emails: string[];
+};
